requirements.txt
flask
werkzeug
gunicorn
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask-assets
cssmin
webassets